requirements.txt
keras-tuner==1.4.7
matplotlib==3.8.4
mediapipe==0.10.14
numpy
opencv-python-headless==4.9.0.80
pandas==2.2.2
plotly==5.22.0
scikit-learn==1.2.2
seaborn==0.13.2
streamlit==1.33.0
tensorflow==2.13.0
tqdm==4.66.2